feat: add sql trait
