+++ classes/local/step/reader_sql.php
@@ -39,6 +39,7 @@ use tool_dataflows\parser;
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class reader_sql extends reader_step {
+    use sql_trait;
 
     /**
      * Return the definition of the fields available in this form.
@@ -182,112 +183,9 @@ class reader_sql extends reader_step {
         }
 
         // Evalulate any remaining expressions as per normal.
-        // NOTE: The expression statement itself MUST be on a single line (currently).
-        $hasexpression = true;
-        $max = 5;
-        while ($hasexpression && $max) {
-            $finalsql = $variables->evaluate($finalsql, function ($message, $e) {
-                // Process the message and clarify it if required.
-                $message = $this->clarify_parser_error($message);
-
-                // Log the message.
-                $this->enginestep->log($message);
-
-                // Throw the original exception (i.e. for the real stack trace).
-                throw $e;
-            });
-            if (!is_string($finalsql)) {
-                throw new \moodle_exception('reader_sql:finalsql_not_string', 'tool_dataflows');
-            }
-            [$hasexpression] = $parser->has_expression($finalsql);
-            $max--;
-        }
+        $finalsql = $this->evaluate_expressions($finalsql);
 
         return $finalsql;
-    }
-
-    /**
-     * Returns a clarified error message if applicable.
-     *
-     * @param   string $message
-     * @param   string $sql replacement for the expression held by default.
-     * @return  string clarified message if applicable
-     */
-    private function clarify_parser_error(string $message, ?string $sql = null): string {
-        // Check and massage the message if required.
-        $matches = null;
-        preg_match_all(
-            // phpcs:disable moodle.Strings.ForbiddenStrings.Found
-            '/Variable "(?<expressionpath>.*)" is not valid.*`(?<expression>.*)`.*for expression (?<sql>.*)/ms',
-            $message,
-            $matches,
-            PREG_SET_ORDER);
-
-        if (!empty($matches)) {
-            // Modify the SQL, adding a pointer to the first instance of the usage which resulted in the error.
-            $match = (object) reset($matches);
-
-            // Replace the field (using the sql key) in the matching expression with the one provided.
-            if (isset($sql)) {
-                $match->sql = $sql;
-            }
-
-            // Pinpoint the line and column of the expression in the SQL.
-            $line = 0;
-            $column = 0;
-            $sqlbylines = explode("\n", $match->sql);
-            foreach ($sqlbylines as $lineindex => $linecontents) {
-                $position = strpos($linecontents, $match->expression);
-                if ($position !== false) {
-                    $column = $position + 1;
-                    $line = $lineindex + 1;
-                    break;
-                }
-            }
-            // Insert the characters in the following line.
-            $sqlwithannotations = $this->draw_arrow_to_string_position($match->sql, $column, $line);
-            $a = (object) array_merge((array) $match, [
-                'sql' => $sqlwithannotations,
-                'column' => $column,
-                'line' => $line,
-            ]);
-            $message = get_string('reader_sql:variable_not_valid_in_position_replacement_text', 'tool_dataflows', $a);
-        }
-
-        return $message;
-    }
-
-    /**
-     * Draws an arrow pointing at the focused position in a string
-     *                                ^
-     * @param   string $string of the original contents
-     * @param   int $column column position to focus on
-     * @param   int $line line position to focus on
-     * @return  string with arrow annotation
-     */
-    private function draw_arrow_to_string_position(string $string, int $column, int $line) {
-        // Split by lines to ensure we insert it at the appropriate line.
-        $sqlbylines = explode("\n", $string);
-
-        // Insert the characters in the following line.
-        $arrow = str_pad('', $column - 1, ' ') . '^';
-        array_splice($sqlbylines, $line, 0, $arrow);
-
-        return implode("\n", $sqlbylines);
-    }
-
-    /**
-     * Validate the configuration settings.
-     *
-     * @param   object $config
-     * @return  true|\lang_string[] true if valid, an array of errors otherwise
-     */
-    public function validate_config($config) {
-        $errors = [];
-        if (empty($config->sql)) {
-            $errors['config_sql'] = get_string('config_field_missing', 'tool_dataflows', 'sql', true);
-        }
-        return empty($errors) ? true : $errors;
     }
 
     /**
